+++ packages/@webex/plugin-meetings/test/unit/spec/breakouts/breakout.ts
@@ -7,6 +7,8 @@ import Metrics from '@webex/plugin-meetings/src/metrics';
 import sinon from 'sinon';
 import {eventType} from '../../../../src/metrics/config';
 import uuid from 'uuid';
+import breakoutEvent from "../../../../src/breakouts/events";
+import SelfUtils from "../../../../src/locus-info/selfUtils";
 describe('plugin-meetings', () => {
   describe('breakout', () => {
     let webex;
@@ -30,6 +32,9 @@ describe('plugin-meetings', () => {
         },
       };
       webex.request = sinon.stub().returns(Promise.resolve('REQUEST_RETURN_VALUE'));
+      webex.meetings = {};
+      webex.meetings.getMeetingByType = sinon.stub();
+
     });
 
     describe('initialize', () => {
@@ -40,12 +45,14 @@ describe('plugin-meetings', () => {
 
     describe('#join', () => {
       it('makes the request as expected', async () => {
-        Metrics.postEvent = sinon.stub();
+        uuid.v4 = sinon.stub().returns('breakoutMoveId');
         const result = await breakout.join();
         assert.calledOnceWithExactly(webex.request, {
           method: 'POST',
           uri: 'url/move',
           body: {
+            breakoutMoveId: 'breakoutMoveId',
+            deviceUrl: undefined,
             groupId: 'groupId',
             sessionId: 'sessionId',
           },
@@ -53,29 +60,28 @@ describe('plugin-meetings', () => {
 
         assert.equal(result, 'REQUEST_RETURN_VALUE');
       });
+
       it('send metrics as expected', async () => {
-        Metrics.postEvent = sinon.stub();
+        breakout.webex.internal.device.url = 'device-url';
+        breakout.webex.meetings = {
+          getMeetingByType: sinon.stub().returns({
+          id: 'meeting-id'
+        })
+        };
+
+        let onBreakoutMoveRequestStub = sinon.stub(breakoutEvent, 'onBreakoutMoveRequest');
+        let onBreakoutMoveResponseStub = sinon.stub(breakoutEvent, 'onBreakoutMoveResponse');
         uuid.v4 = sinon.stub().returns('breakoutMoveId');
         await breakout.join();
-        assert.calledTwice(Metrics.postEvent);
-        assert.calledWithMatch(Metrics.postEvent, {
-          event: eventType.MOVE_TO_BREAKOUT,
-          meetingId: 'activeMeetingId',
-          data: {
-            breakoutMoveId: 'breakoutMoveId',
-            breakoutSessionId: 'sessionId',
-            breakoutGroupId: 'groupId',
-          },
-        });
-        assert.calledWithMatch((Metrics.postEvent as any).secondCall, {
-          event: eventType.JOIN_BREAKOUT_RESPONSE,
-          meetingId: 'activeMeetingId',
-          data: {
-            breakoutMoveId: 'breakoutMoveId',
-            breakoutSessionId: 'sessionId',
-            breakoutGroupId: 'groupId',
-          },
-        });
+        assert.calledOnceWithExactly(breakoutEvent.onBreakoutMoveRequest,
+          {currentSession: breakout, meeting: {id: 'meeting-id'}, breakoutMoveId: 'breakoutMoveId'}
+        );
+        assert.calledOnceWithExactly(breakoutEvent.onBreakoutMoveResponse,
+          {currentSession: breakout, meeting: {id: 'meeting-id'}, breakoutMoveId: 'breakoutMoveId'}
+        );
+
+        onBreakoutMoveRequestStub.restore();
+        onBreakoutMoveResponseStub.restore();
       });
     });